+++ dsmr_datalogger/tests/datalogger/test_generic.py
@@ -1,3 +1,5 @@
+from unittest import mock
+
 from django.core.management import CommandError
 from django.test import TestCase
 from django.utils import timezone
@@ -7,6 +9,7 @@ import pytz
 from dsmr_backend.tests.mixins import InterceptStdoutMixin
 from dsmr_datalogger.models.settings import DataloggerSettings
 from dsmr_datalogger.models.reading import DsmrReading, MeterStatistics
+from dsmr_datalogger.exceptions import InvalidTelegramChecksum
 import dsmr_datalogger.services
 
 
@@ -105,7 +108,7 @@ class TestServices(TestCase):
             "0-1:96.1.0(xxxxxxxxxxxxx)\r\n",
             "0-1:24.2.1(151110190000W)(00845.206*m3)\r\n",
             "0-1:24.4.0(1)\r\n",
-            "!D19A\n",
+            "!AF0C\n",
         ])
 
         self.assertIsNone(MeterStatistics.get_solo().electricity_tariff)  # Empty model in DB.
@@ -128,14 +131,15 @@ class TestServices(TestCase):
         self.assertEqual(meter_statistics.voltage_sag_count_l1, 2)
         self.assertEqual(meter_statistics.voltage_sag_count_l2, 2)
 
-    def test_extra_devices_mbus_hack(self):
+    @mock.patch('dsmr_datalogger.services.verify_telegram_checksum')
+    def test_extra_devices_mbus_hack(self, *mocks):
         """ Verify that the hack issue #92 is working. """
         fake_telegram = [
             "/XMX5LGBBFFB123456789\r\n",
             "\r\n",
             "1-3:0.2.8(40)\r\n",
             "0-0:1.0.0(151110192959W)\r\n",
-            "0-0:96.1.1(xxxxxxxxxxxxx)\r\n",
+            "0-0:96.1.1(123456789012345678901234567890)\r\n",
             "1-0:1.8.1(000510.747*kWh)\r\n",
             "1-0:2.8.1(000000.123*kWh)\r\n",
             "1-0:1.8.2(000500.013*kWh)\r\n",
@@ -154,10 +158,10 @@ class TestServices(TestCase):
             "1-0:32.36.0(00000)\r\n",
             "1-0:52.36.0(00000)\r\n",
             "1-0:72.36.0(00000)\r\n",
-            "0-1:96.1.0(xxxxxxxxxxxxx)\r\n",
+            "0-1:96.1.0(098765432109876543210987654321)\r\n",
             "0-1:24.2.1(151110190000W)(00845.206*m3)\r\n",
             "0-1:24.4.0(1)\r\n",
-            "!D19A\n",
+            "!6013\n",
         ]
         self.assertFalse(DsmrReading.objects.exists())
         dsmr_datalogger.services.telegram_to_reading(data=''.join(fake_telegram))
@@ -170,6 +174,57 @@ class TestServices(TestCase):
 
             dsmr_datalogger.services.telegram_to_reading(data=''.join(fake_telegram))
             self.assertFalse(DsmrReading.objects.filter(extra_device_timestamp__isnull=True).exists())
+
+    def test_verify_telegram_checksum(self):
+        """ Verify that CRC checks. """
+        telegram = [
+            "/XMX5LGBBFFB123456789\r\n",
+            "\r\n",
+            "1-3:0.2.8(40)\r\n",
+            "0-0:1.0.0(151110192959W)\r\n",
+            "0-0:96.1.1(123456789012345678901234567890)\r\n",
+            "1-0:1.8.1(000510.747*kWh)\r\n",
+            "1-0:2.8.1(000000.123*kWh)\r\n",
+            "1-0:1.8.2(000500.013*kWh)\r\n",
+            "1-0:2.8.2(000123.456*kWh)\r\n",
+            "0-0:96.14.0(0001)\r\n",
+            "1-0:1.7.0(00.192*kW)\r\n",
+            "1-0:2.7.0(00.123*kW)\r\n",
+            "0-0:17.0.0(999.9*kW)\r\n",
+            "0-0:96.3.10(1)\r\n",
+            "0-0:96.7.21(00003)\r\n",
+            "0-0:96.7.9(00000)\r\n",
+            "1-0:99.97.0(0)(0-0:96.7.19)\r\n",
+            "1-0:32.32.0(00002)\r\n",
+            "1-0:52.32.0(00002)\r\n",
+            "1-0:72.32.0(00000)\r\n",
+            "1-0:32.36.0(00000)\r\n",
+            "1-0:52.36.0(00000)\r\n",
+            "1-0:72.36.0(00000)\r\n",
+            "0-0:96.13.1()\r\n",
+            "0-0:96.13.0()\r\n",
+            "1-0:31.7.0(000*A)\r\n",
+            "1-0:51.7.0(000*A)\r\n",
+            "1-0:71.7.0(001*A)\r\n",
+            "1-0:21.7.0(00.000*kW)\r\n",
+            "1-0:41.7.0(00.000*kW)\r\n",
+            "1-0:61.7.0(00.192*kW)\r\n",
+            "1-0:22.7.0(00.000*kW)\r\n",
+            "1-0:42.7.0(00.000*kW)\r\n",
+            "1-0:62.7.0(00.000*kW)\r\n",
+            "0-1:24.1.0(003)\r\n",
+            "0-1:96.1.0(098765432109876543210987654321)\r\n",
+            "0-1:24.2.1(151110190000W)(00845.206*m3)\r\n",
+            "0-1:24.4.0(1)\r\n",
+            "!D19A\n",
+        ]
+
+        with self.assertRaises(InvalidTelegramChecksum):
+            dsmr_datalogger.services.verify_telegram_checksum(data=''.join(telegram))
+
+        # Again, with the correct checksum.
+        telegram[-1] = "!58C8\n"
+        dsmr_datalogger.services.verify_telegram_checksum(data=''.join(telegram))
 
 
 class TestDsmrVersionMapping(InterceptStdoutMixin, TestCase):